+++ libcrux-ml-dsa/src/types.rs
@@ -86,3 +86,83 @@ pub enum SigningError {
     RejectionSamplingError,
     ContextTooLongError,
 }
+
+#[cfg(feature = "codec")]
+mod codec {
+    use super::*;
+
+    macro_rules! impl_tls_codec_for_generic_struct {
+        ($name:ident) => {
+            // XXX: `tls_codec::{Serialize, Deserialize}` are only
+            // available for feature `std`. For `no_std` scenarios, we
+            // need to implement `tls_codec::{SerializeBytes,
+            // DeserializeBytes}`, but `SerializeBytes` is not
+            // implemented for `VLByteSlice`.
+            impl<const SIZE: usize> tls_codec::DeserializeBytes for $name<SIZE> {
+                fn tls_deserialize_bytes(bytes: &[u8]) -> Result<(Self, &[u8]), tls_codec::Error> {
+                    let (bytes, remainder) = tls_codec::VLBytes::tls_deserialize_bytes(bytes)?;
+                    Ok((
+                        Self {
+                            value: bytes
+                                .as_ref()
+                                .try_into()
+                                .map_err(|_| tls_codec::Error::InvalidInput)?,
+                        },
+                        remainder,
+                    ))
+                }
+            }
+
+            #[cfg(feature = "std")]
+            impl<const SIZE: usize> tls_codec::Serialize for $name<SIZE> {
+                fn tls_serialize<W: std::io::Write>(
+                    &self,
+                    writer: &mut W,
+                ) -> Result<usize, tls_codec::Error> {
+                    let out = tls_codec::VLByteSlice(self.as_ref());
+                    out.tls_serialize(writer)
+                }
+            }
+
+            #[cfg(feature = "std")]
+            impl<const SIZE: usize> tls_codec::Serialize for &$name<SIZE> {
+                fn tls_serialize<W: std::io::Write>(
+                    &self,
+                    writer: &mut W,
+                ) -> Result<usize, tls_codec::Error> {
+                    (*self).tls_serialize(writer)
+                }
+            }
+
+            #[cfg(feature = "std")]
+            impl<const SIZE: usize> tls_codec::Deserialize for $name<SIZE> {
+                fn tls_deserialize<R: std::io::Read>(
+                    bytes: &mut R,
+                ) -> Result<Self, tls_codec::Error> {
+                    let bytes = tls_codec::VLBytes::tls_deserialize(bytes)?;
+                    Ok(Self {
+                        value: bytes
+                            .as_ref()
+                            .try_into()
+                            .map_err(|_| tls_codec::Error::InvalidInput)?,
+                    })
+                }
+            }
+
+            impl<const SIZE: usize> tls_codec::Size for $name<SIZE> {
+                fn tls_serialized_len(&self) -> usize {
+                    tls_codec::VLByteSlice(self.as_ref()).tls_serialized_len()
+                }
+            }
+
+            impl<const SIZE: usize> tls_codec::Size for &$name<SIZE> {
+                fn tls_serialized_len(&self) -> usize {
+                    (*self).tls_serialized_len()
+                }
+            }
+        };
+    }
+
+    impl_tls_codec_for_generic_struct!(MLDSAVerificationKey);
+    impl_tls_codec_for_generic_struct!(MLDSASignature);
+}